+++ controllers/index.js
@@ -5,6 +5,8 @@ const transactionManageController = require('./transactionManageController')
 const transactionController = require("./transactionController")
 const ongkirController = require("./ongkirController")
 const financialReportController = require("./financialReportController")
+const productController = require("./productController")
+const parcelController = require("./parcelController")
 
 module.exports = {
     userController,
@@ -13,5 +15,7 @@ module.exports = {
     transactionManageController,
     transactionController,
     ongkirController,
-    financialReportController
+    financialReportController,
+    productController,
+    parcelController
 }
