+++ index.js
@@ -16,7 +16,9 @@ const {
     transactionManageRouter, 
     transactionRouter, 
     ongkirRouter, 
-    financialReportRouter 
+    financialReportRouter,
+    productRouter,
+    parcelRouter
 } = require('./routers')
 
 app.use(cors()) // get data from front-end
@@ -25,6 +27,8 @@ app.use('/static', express.static(path.join(__dirname, 'public'))) // access sta
 app.use(bearerToken()) // read token 
 app.use('/profile', profileRouter)
 app.use('/auth', userRouter)
+app.use('/product', productRouter)
+app.use('/parcel', parcelRouter)
 app.use('/product-manage', productManageRouter)
 app.use('/transaction-manage', transactionManageRouter)
 app.use('/financial-report', financialReportRouter)
